Boton deshacer al eliminar un negocio de favoritos

+++ app/src/main/java/com/fdez/projecttfg/adapters/NegocioAdapter.kt
@@ -94,25 +94,41 @@ class NegocioAdapter(
 
                 }
                 override fun unLiked(likeButton: LikeButton) {
-                    // Elimina el documento de la colección "likes" que corresponde a este usuario y a este negocio
-                    val alias = negocio.alias
-                    val userId = FirebaseAuth.getInstance().currentUser?.uid
-                    val query = offersCollection.whereEqualTo("id_user", userId).whereEqualTo("alias", alias)
-                    query.get().addOnSuccessListener { documents ->
-                        for (document in documents) {
-                            document.reference.delete()
-                                .addOnSuccessListener {
-                                    Log.d(TAG, "Documento eliminado correctamente")
-                                }
-                                .addOnFailureListener { e ->
-                                    Log.w(TAG, "Error al eliminar el documento", e)
-                                }
-                        }
-                    }.addOnFailureListener { exception ->
-                        Log.w(TAG, "Error al obtener el documento: ", exception)
+                    // Elimina el negocio de la lista y notifica al adaptador sobre el cambio
+                    val position = adapterPosition
+                    if (position != RecyclerView.NO_POSITION) {
+                        (negocios as MutableList<Negocio>).remove(negocio)
+                        notifyItemRemoved(position)
                     }
-
-
+                    val snackbar = Snackbar.make(likeButton, "Negocio eliminado", Snackbar.LENGTH_LONG)
+                    snackbar.setAction("Deshacer") {
+                        (negocios as MutableList<Negocio>).add(position, negocio)
+                        notifyItemInserted(position)
+                    }
+                    snackbar.addCallback(object : Snackbar.Callback() {
+                        override fun onDismissed(transientBottomBar: Snackbar?, event: Int) {
+                            if (event != Snackbar.Callback.DISMISS_EVENT_ACTION) {
+                                // El Snackbar se ha descartado sin seleccionar "Deshacer", por lo tanto, eliminamos el documento
+                                val alias = negocio.alias
+                                val userId = FirebaseAuth.getInstance().currentUser?.uid
+                                val query = offersCollection.whereEqualTo("id_user", userId).whereEqualTo("alias", alias)
+                                query.get().addOnSuccessListener { documents ->
+                                    for (document in documents) {
+                                        document.reference.delete()
+                                            .addOnSuccessListener {
+                                                Log.d(TAG, "Documento eliminado correctamente")
+                                            }
+                                            .addOnFailureListener { e ->
+                                                Log.w(TAG, "Error al eliminar el documento", e)
+                                            }
+                                    }
+                                }.addOnFailureListener { exception ->
+                                    Log.w(TAG, "Error al obtener el documento: ", exception)
+                                }
+                            }
+                        }
+                    })
+                    snackbar.show()
 
                 }
             })
@@ -136,6 +152,14 @@ class NegocioAdapter(
 
         }
 
+    }
+    // Agrega el método removeNegocio() para eliminar un elemento de la lista
+    fun removeNegocio(negocio: Negocio) {
+        val position = negocios.indexOf(negocio)
+        if (position != -1) {
+            (negocios as MutableList<Negocio>).remove(negocio)
+            notifyItemRemoved(position)
+        }
     }
 
     fun setOnItemClickListener(listener: OnItemClickListenerNegocio) {
